requirements.txt
fastapi
uvicorn
torch
pillow
transformers
python-multipart
